+++ src/main/java/net/congstar/jira/plugins/scrumpoker/action/ChooseCardAction.java
@@ -1,5 +1,7 @@
 package net.congstar.jira.plugins.scrumpoker.action;
 
+import com.atlassian.jira.security.JiraAuthenticationContext;
+import com.atlassian.jira.web.HttpServletVariables;
 import net.congstar.jira.plugins.scrumpoker.data.ScrumPokerStorage;
 
 /**
@@ -8,19 +10,25 @@ import net.congstar.jira.plugins.scrumpoker.data.ScrumPokerStorage;
 public class ChooseCardAction extends ScrumPokerAction {
 
     private static final long serialVersionUID = 1L;
-    private static final String PARAM_CHOSEN_CARD = "chosenCard";
+    static final String PARAM_CHOSEN_CARD = "chosenCard";
 
     private final ScrumPokerStorage scrumPokerStorage;
+    private final JiraAuthenticationContext jiraAuthenticationContext;
+    private final HttpServletVariables httpServletVariables;
 
-    public ChooseCardAction(ScrumPokerStorage scrumPokerStorage) {
+    public ChooseCardAction(ScrumPokerStorage scrumPokerStorage, JiraAuthenticationContext jiraAuthenticationContext,
+                            HttpServletVariables httpServletVariables) {
         this.scrumPokerStorage = scrumPokerStorage;
+        this.jiraAuthenticationContext = jiraAuthenticationContext;
+        this.httpServletVariables = httpServletVariables;
     }
 
     @Override
     protected String doExecute() throws Exception {
-        String issueKey = getHttpRequest().getParameter(PARAM_ISSUE_KEY);
-        String chosenCard = getHttpRequest().getParameter(PARAM_CHOSEN_CARD);
-        scrumPokerStorage.sessionForIssue(issueKey, currentUserKey()).updateCard(currentUserKey(), chosenCard);
+        String issueKey = httpServletVariables.getHttpRequest().getParameter(PARAM_ISSUE_KEY);
+        String chosenCard = httpServletVariables.getHttpRequest().getParameter(PARAM_CHOSEN_CARD);
+        String currentUserKey = jiraAuthenticationContext.getLoggedInUser().getKey();
+        scrumPokerStorage.sessionForIssue(issueKey, currentUserKey).updateCard(currentUserKey, chosenCard);
         return openScrumPokerForIssue(issueKey);
     }